api/requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.4 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
asgiref==3.6.0 ; python_version >= '3.7'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==23.1.0 ; python_version >= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
contourpy==1.0.7 ; python_version >= '3.8'
cycler==0.11.0 ; python_version >= '3.6'
django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
dnspython==2.3.0 ; python_version >= '3.7' and python_version < '4.0'
fonttools==4.39.3 ; python_version >= '3.8'
frozenlist==1.3.3 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
joblib==1.2.0 ; python_version >= '3.7'
kiwisolver==1.4.4 ; python_version >= '3.7'
loguru==0.7.0 ; python_version >= '3.5'
matplotlib==3.7.1
multidict==6.0.4 ; python_version >= '3.7'
numpy==1.24.3 ; python_version >= '3.8'
openai==0.27.5
packaging==23.1 ; python_version >= '3.7'
pandas==2.0.1
pillow==9.5.0 ; python_version >= '3.7'
pinecone-client==2.2.1
plotly==5.14.1
pyparsing==3.0.9 ; python_full_version >= '3.6.8'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
pytz==2023.3
pyyaml==6.0 ; python_version >= '3.6'
requests==2.29.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4 ; python_version >= '3.5'
tenacity==8.2.2 ; python_version >= '3.6'
threadpoolctl==3.1.0 ; python_version >= '3.6'
tqdm==4.65.0 ; python_version >= '3.7'
typing-extensions==4.5.0 ; python_version >= '3.7'
tzdata==2023.3 ; python_version >= '2'
urllib3==1.26.15 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
whitenoise==6.4.0
yarl==1.9.2 ; python_version >= '3.7'
